refactor(account-set): align batch validation
The first batch implementation placed graph orchestration in the
repository, unlike account attachment where the service locks, the graph
cache validates, and the repository persists. That split made ownership
of graph reads and invariants inconsistent between the two paths.

Route batch validation through SetGraphCache after the service acquires
the structure lock. Keep SQL loading and persistence in AccountSetRepo,
so both membership APIs follow the same lock, validate, persist sequence
without changing batch behavior.
